requirements.txt
openai>=1.12.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
sounddevice>=0.4.6
numpy>=1.24.0
pydub>=0.25.1
simpleaudio>=1.0.4
python-dateutil>=2.8.2
pytz>=2024.1

